wip: refactor all codes
