--- Form/connector.php
@@ -1,11 +1,0 @@
-<?php
-
-$dbhost = "localhost";
-$dbuser = "root";
-$dbpass = "";
-$dbname = "uc";
-
-if (!$con =mysqli_connect ($dbhost, $dbuser, $dbpass, $dbname))
-{
-	die ("failed to connect!");
-}	?>
