+++ admin/db.php
@@ -1,0 +1,13 @@
+<?php
+$servername = "localhost";
+$username = "root";
+$password = "";
+$database = "labuser";
+
+$connection = new mysqli($servername, $username, $password, $database);
+if ($connection === false) {
+    die("connection error");
+}
+
+
+?>
